Rename `phases.py` to `forks.py`; add `ValueError` check
