+++ tests/core/pyspec/eth2spec/test/altair/unittests/test_config_override.py
@@ -5,7 +5,7 @@ from eth2spec.test.context import (
     with_phases,
 )
 from eth2spec.test.helpers.constants import ALTAIR
-from eth2spec.test.helpers.phases import (
+from eth2spec.test.helpers.forks import (
     is_post_capella, is_post_eip4844,
 )
