+++ tests/core/pyspec/eth2spec/test/helpers/attestations.py
@@ -5,8 +5,8 @@ from typing import List
 from eth2spec.test.context import expect_assertion_error
 from eth2spec.test.helpers.state import state_transition_and_sign_block, next_epoch, next_slot
 from eth2spec.test.helpers.block import build_empty_block_for_next_slot
+from eth2spec.test.helpers.forks import is_post_altair
 from eth2spec.test.helpers.keys import privkeys
-from eth2spec.test.helpers.phases import is_post_altair
 from eth2spec.utils import bls
 from eth2spec.utils.ssz.ssz_typing import Bitlist
