+++ tests/core/pyspec/eth2spec/test/helpers/block.py
@@ -1,6 +1,6 @@
 from eth2spec.test.helpers.execution_payload import build_empty_execution_payload
+from eth2spec.test.helpers.forks import is_post_altair, is_post_bellatrix
 from eth2spec.test.helpers.keys import privkeys
-from eth2spec.test.helpers.phases import is_post_altair, is_post_bellatrix
 from eth2spec.utils import bls
 from eth2spec.utils.bls import only_with_bls
 from eth2spec.utils.ssz.ssz_impl import hash_tree_root
