+++ tests/core/pyspec/eth2spec/test/helpers/deposits.py
@@ -1,8 +1,8 @@
 from random import Random
 
 from eth2spec.test.context import expect_assertion_error
+from eth2spec.test.helpers.forks import is_post_altair
 from eth2spec.test.helpers.keys import pubkeys, privkeys
-from eth2spec.test.helpers.phases import is_post_altair
 from eth2spec.test.helpers.state import get_balance
 from eth2spec.utils import bls
 from eth2spec.utils.merkle_minimal import calc_merkle_tree_from_leaves, get_merkle_proof
