+++ tests/core/pyspec/eth2spec/test/helpers/random.py
@@ -2,7 +2,7 @@ from random import Random
 
 from eth2spec.test.helpers.attestations import cached_prepare_state_with_attestations
 from eth2spec.test.helpers.deposits import mock_deposit
-from eth2spec.test.helpers.phases import is_post_altair
+from eth2spec.test.helpers.forks import is_post_altair
 from eth2spec.test.helpers.state import next_epoch
