+++ tests/core/pyspec/eth2spec/test/phase0/epoch_processing/test_process_justification_and_finalization.py
@@ -3,7 +3,7 @@ from eth2spec.test.context import spec_state_test, with_all_phases
 from eth2spec.test.helpers.epoch_processing import (
     run_epoch_processing_with,
 )
-from eth2spec.test.helpers.phases import is_post_altair
+from eth2spec.test.helpers.forks import is_post_altair
 from eth2spec.test.helpers.state import transition_to, next_epoch_via_block, next_slot
 from eth2spec.test.helpers.voluntary_exits import get_unslashed_exited_validators
