+++ tests/core/pyspec/eth2spec/test/phase0/genesis/test_validity.py
@@ -8,7 +8,7 @@ from eth2spec.test.helpers.constants import MINIMAL
 from eth2spec.test.helpers.deposits import (
     prepare_full_genesis_deposits,
 )
-from eth2spec.test.helpers.phases import (
+from eth2spec.test.helpers.forks import (
     is_post_altair,
 )
